[Backend] Clean up AxisInfo (NFC).
- Remove unused member variables.
- Clarify and reformat comments.
- Move code out of header.
- Add namespaces.

GPC: axis-info-cleanup
